requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
#mysqlclient==2.2.1
psycopg2-binary
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
gunicorn
psycopg2-binary
django-phonenumber-field==7.0.2
phonenumbers==8.12.51
drf-yasg==1.21.7